app/backend/requirements/base.txt
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
pytest-django==4.1.0
django-allauth==0.44.0
psycopg2-binary==2.8.6
celery==4.4.7
django-celery-beat==2.1.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.6.0